Fixed how publisher sets topic
If no topic/channel is specified, them ZMQTOPICEND end is not inserted.
As a result, the message bus (which is only responsible for forwarding to
any published message to its subscribers), does not prefix a ZMQTOPICEND
to messages it receives (which would otherwise result in 2 of these to
be in the message) and allows for its subscribers to receive
messages by the topic they filter for.